requirements.txt
google-generativeai
python-dotenv
python-docx
Pillow